+++ testEpicsApp/src/uint32DigitalDriver.c
@@ -410,7 +410,7 @@ static const char *testDriverReason = "testDriverReason";
 static const char *skipWhite(const char *pstart)
 {
     const char *p = pstart;
-    while(*p && isspace(*p)) p++;
+    while(*p && isspace((int)*p)) p++;
     return p;
 }
